Always set default context.
Previously if a customer sends the `.context` field, we would not send
any default context fields, such as `context.library`.

The new behvaiour is that a customer can send `context`, and we will
merge it with the default context before sending.

This change fixes this behaviour so that the user defined `context` is
merged with the default Segment context.

The implementation does allow the user to send a custom
`context.library`. I did this because that's what the old behaviour was.

Closes https://github.com/segmentio/analytics-php/issues/88.